remove duplicated implementation of TestStreamInput
